wip(4.1-traveler-profile-type): agent changes
