+++ src/lib/profile/types.ts
@@ -1,0 +1,8 @@
+export type TravelerProfile = {
+  full_name: string | null;
+  avatar_url: string | null;
+  bio: string | null;
+  home_city: string | null;
+  languages: string[] | null;
+  birth_year: number | null;
+};
